test: fold svelte style guard into the plugin test file
